Optimize Jackie Chan

+++ catscripts/Entertainment/Actors_and_actresses/Jackie_Chan/JackieChan.sh
@@ -1,16 +1,19 @@
 #!/bin/bash
 
-export KEYWORDS_ACTOR_JACKIECHAN="Jackie(| )Chan"
+KEYWORDS_ACTOR_JACKIECHAN="Jackie(| )Chan"
 
-egrep -i "$KEYWORDS_ACTOR_JACKIECHAN" newpages.txt >> JackieChan.txt
-
-JACKIECHAN=`stat --print=%s JackieChan.txt`
-
-if [ $JACKIECHAN -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="JackieChan.txt"
-  export CATNAME="Jackie Chan"
-  $CATEGORIZE
-fi
 
-rm JackieChan.txt
+  JACKIECHAN=`egrep -i "$KEYWORDS_ACTOR_JACKIECHAN" newpages.txt`
+
+  if [ $JACKIECHAN -ne 0 ];
+  then
+    egrep -i "$KEYWORDS_ACTOR_JACKIECHAN" newpages.txt > JackieChan.txt
+    export CATFILE="JackieChan.txt"
+    export CATNAME="Jackie Chan"
+    $CATEGORIZE
+    rm JackieChan.txt
+    unset JACKIECHAN
+  fi
+fi
